+++ include/tvm/relay/attrs/memory.h
@@ -32,6 +32,30 @@ namespace tvm {
 namespace relay {
 
 /*!
+ * \brief Options for allocating storage.
+ */
+struct AllocStorageAttrs : public tvm::AttrsNode<AllocStorageAttrs> {
+  DataType dtype;
+  int device_id;
+  int device_type;
+
+  TVM_DECLARE_ATTRS(AllocStorageAttrs, "relay.attrs.AllocStorageAttrs") {
+    TVM_ATTR_FIELD(dtype)
+      .describe(
+         "The dtype of the tensor to allocate.")
+      .set_default(DataType::Float(32, 1));
+    TVM_ATTR_FIELD(device_id)
+      .describe(
+        "The device id on which to allocate memory."
+      );
+    TVM_ATTR_FIELD(device_type)
+      .describe(
+        "The device type on which to allocate memory."
+      );
+  }
+};
+
+/*!
  * \brief Options for allocating tensors.
  */
 struct AllocTensorAttrs : public tvm::AttrsNode<AllocTensorAttrs> {
